chore: add tests generation
