+++ codegen/Cargo.toml
@@ -17,3 +17,4 @@ clap = { version = "3.1", default-features = false, features = ["derive", "std"]
 toml = "0.5"
 serde = { version = "1.0", features = ["derive"] }
 prettyplease = "0.1"
+heck = "0.4"
