--- v_escape/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "v_escape"
-version = "0.18.0"
-authors = ["Juan Aguilar Santillana <mhpoin@gmail.com>"]
-description = "The simd optimized escaping code"
-documentation = "https://docs.rs/v_escape"
-edition = "2021"
-keywords = ["escaping", "simd"]
-license = "MIT/Apache-2.0"
-readme = "../README.md"
-repository = "https://github.com/botika/v_escape"
-workspace = ".."
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-[dependencies]
-v_escape_derive = { version = "0.9", path = "../v_escape_derive" }
-buf-min = { version = "0.6.0", default-features = false  }
